+++ .gitignore
@@ -7,4 +7,4 @@ cmake-*/
 *.diff
 html/
 debug.log
-benchmarks/*.csv
+benchmark/*.csv
